Rename `search.rs` to `search_recipe.rs` for consistency
