+++ raphael-cli/src/commands/mod.rs
@@ -1,4 +1,4 @@
 pub mod ingredients;
-pub mod search;
+pub mod search_recipe;
 pub mod search_mission;
 pub mod solve;
