+++ raphael-cli/src/commands/search_recipe.rs
